% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_names.R
\name{change_flagged_names}
\alias{change_flagged_names}
\title{Title}
\usage{
change_flagged_names(data, names = ".name", species = "species")
}
\arguments{
\item{data}{new dataset}

\item{names}{name of column that indicates which observations need a name change}

\item{species}{species column name in dataset}
}
\description{
Title
}
