% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull_flags.R
\name{check_convex_hull}
\alias{check_convex_hull}
\title{Function to check if observations of specific species are within a convex hull
of species on the same continent}
\usage{
check_convex_hull(
  data,
  ECD,
  ECDContinent = "continent",
  datContinent = "continent",
  ECDspecies = "species",
  datSpecies = "species",
  ECDlon = "decimalLongitude",
  ECDlat = "decimalLatitude",
  datLon = "decimalLongitude",
  datLat = "decimalLatitude"
)
}
\arguments{
\item{data}{dataset of observations}

\item{ECD}{Expert curated dataset}

\item{ECDContinent}{Continent column name in ECD}

\item{datContinent}{Continent column name in new dataset}

\item{ECDspecies}{Species column name in ECD}

\item{datSpecies}{Species column name in new dataset}

\item{ECDlon}{Longitude column name in ECD}

\item{ECDlat}{Latitude column name in ECD}

\item{datLon}{Longitude column name in new dataset}

\item{datLat}{Latitude column name in new dataset}
}
\value{
dataset of the same class passed to it. New column of convex hull flags.
}
\description{
Function to check if observations of specific species are within a convex hull
of species on the same continent
}
\examples{
check_convex_hull(obsData, ECD)
}
