% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data.R
\name{combine}
\alias{combine}
\title{Function to combine new data and ECD into one dataframe}
\usage{
combine(data, ECD)
}
\arguments{
\item{data}{dataset with new occurence data}

\item{ECD}{Expert curated dataset}
}
\value{
dataframe
}
\description{
Function to combine new data and ECD into one dataframe
}
