% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pseudobulk.R
\name{create_pseudo_bulk}
\alias{create_pseudo_bulk}
\title{Create pseudobulks}
\usage{
create_pseudo_bulk(expr, meta)
}
\arguments{
\item{expr}{Expression matrix where columns are cell IDs and rows are gene names.}

\item{meta}{Dataframe where each row should represent a cell. Cell IDs should be the rownames. It needs to have the column "patientID" and "disorder". "disorder" should have the values "yes" or "no".}
}
\value{
A list containing the pseudobulk matrix and its metadata.
}
\description{
Create pseudobulks
}
