% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cpm_log.R
\name{do_cpm_log}
\alias{do_cpm_log}
\title{CPM normalizes a matrix and log-transforms it}
\usage{
do_cpm_log(mtx, log = FALSE)
}
\arguments{
\item{mtx}{A matrix.}

\item{log}{TRUE or FALSE. Whether to log-transform the matrix as well. Default is false.}
}
\value{
A CPM normalized matrix.
}
\description{
CPM normalizes a matrix and log-transforms it
}
\examples{
#' set.seed(123)
ctmat <- matrix(sample(c(0, 1, 2, 3), 30, replace = TRUE),
nrow = 5, ncol = 6,
dimnames = list(
paste0("Gene", 1:5),
paste0("Cell", 1:6)))
do_cpm_log(ctmat, log = TRUE)
}
