% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_flags.R
\name{.flag_countries_helper}
\alias{.flag_countries_helper}
\title{Helper function to flag occurences where species name is not found in continent in the Expert
curated dataset}
\usage{
.flag_countries_helper(
  data,
  ECD,
  ECDspecies,
  datSpecies,
  ECDcountry,
  datCountry
)
}
\arguments{
\item{data}{Dataset with observations}

\item{ECD}{Expert curated dataset}

\item{ECDspecies}{Species column name in ECD}

\item{datSpecies}{Species column name in  dataset with observations}

\item{ECDcountry}{Country column name in ECD}

\item{datCountry}{Country column name in dataset with observations}
}
\value{
dataset with country flags column
}
\description{
Helper function to flag occurences where species name is not found in continent in the Expert
curated dataset
}
\examples{
 flag_countries(data, ECD)
}
