% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_flags.R
\name{.flag_distance_helper}
\alias{.flag_distance_helper}
\title{Helper function to flag species occurences that are over a certain distance from the center
of their distribution on a given continet}
\usage{
.flag_distance_helper(
  data,
  ECD,
  ECDContinent,
  datContinent,
  ECDspecies,
  datSpecies,
  ECDlongitude,
  ECDlatitude,
  datLongitude,
  datLatitude
)
}
\arguments{
\item{data}{new dataset}

\item{ECD}{expert curated dataset}

\item{ECDContinent}{continent column name in ECD}

\item{datContinent}{continent column name in new dataset}

\item{ECDspecies}{species column name in ECD}

\item{datSpecies}{species column name in new dataset}

\item{ECDlongitude}{longitude column name in ECD}

\item{ECDlatitude}{latitude column name in ECD}

\item{datLongitude}{longtiude column name in new dataset}

\item{datLatitude}{latitude column name in new dataset}
}
\value{
dataset of the same class passed to it. New column of distance flags
}
\description{
Helper function to flag species occurences that are over a certain distance from the center
of their distribution on a given continet
}
