% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECD}
\alias{ECD}
\title{50,000 observations from expert curated dataset}
\format{
A data frame with 985 rows and 148 variables:
\describe{
\item{scientificNameORpreviousID}{scientific name from gbif}
...
}
}
\source{
\url{https://www.gbif.org/occurrence/search}
}
\usage{
ECD
}
\description{
A dataset containing biological and location data for 50,000 expert curated observations
}
\keyword{datasets}
