% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_flags.R
\name{flag_countries}
\alias{flag_countries}
\title{FLags occurence if species is in a country it does not occur in in ECD}
\usage{
flag_countries(
  data,
  ECD,
  datSpecies = "species",
  ECDspecies = "species",
  ECDcountry = "countryCode",
  datCountry = "countryCode"
)
}
\arguments{
\item{data}{dataset with observations}

\item{ECD}{Expert curated dataset}

\item{datSpecies}{Species column name in  dataset with observations}

\item{ECDspecies}{Species column name in ECD}

\item{ECDcountry}{Country column name in ECD}

\item{datCountry}{Country column name in dataset with observations}
}
\value{
dataset of the same class passed to it. New column of country flags
}
\description{
FLags occurence if species is in a country it does not occur in in ECD
}
\examples{
flag_countries(obsData, ECD)

}
