% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{flag_data}
\alias{flag_data}
\title{flag data using coordinate_cleaner}
\usage{
flag_data(data)
}
\arguments{
\item{data}{biological data to be flagged}
}
\value{
returns a dataframe with included flags
}
\description{
flag data using coordinate_cleaner
}
\examples{
obsData <- data.frame(obsData[1:5])
flag_data(obsData)

}
