% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intoduced_flags.R
\name{flag_deleted}
\alias{flag_deleted}
\title{Flag species labeled as "deleted"}
\usage{
flag_deleted(data, delCol = "databaseChanges", delTitle = "deleted")
}
\arguments{
\item{data}{dataset}

\item{delCol}{name of column that contains introduced species labels}

\item{delTitle}{Label of deleted species}
}
\value{
dataset of the same class passed to it. New column of introduced flags
}
\description{
Flag species labeled as "deleted"
}
