% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_flags.R
\name{flag_distance}
\alias{flag_distance}
\title{Flags species occurences that are over a certain distance from the center of the species distribution}
\usage{
flag_distance(
  data,
  ECD,
  ECDContinent = "continent",
  datContinent = "continent",
  ECDspecies = "species",
  datSpecies = "species",
  ECDlongitude = "decimalLongitude",
  ECDlatitude = "decimalLatitude",
  datLongitude = "decimalLongitude",
  datLatitude = "decimalLatitude"
)
}
\arguments{
\item{data}{new dataset}

\item{ECD}{expert curated dataset}

\item{ECDContinent}{continent column name in ECD}

\item{datContinent}{continent column name in new dataset}

\item{ECDspecies}{species column name in ECD}

\item{datSpecies}{species column name in new dataset}

\item{ECDlongitude}{longitude column name in ECD}

\item{ECDlatitude}{latitude column name in ECD}

\item{datLongitude}{longtiude column name in new dataset}

\item{datLatitude}{latitude column name in new dataset}
}
\value{
dataset of the same class passed to it. New column of distance flags
}
\description{
Flags species occurences that are over a certain distance from the center of the species distribution
}
