% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_names.R
\name{flag_names}
\alias{flag_names}
\title{Flag species names in the new dataset that do not occur in the ECD}
\usage{
flag_names(data, ECD, datSpecies = "species", ECDspecies = "species")
}
\arguments{
\item{data}{new dataset}

\item{ECD}{Expert curated dataset}

\item{datSpecies}{species column name in new dataset}

\item{ECDspecies}{species column name in ECD}
}
\description{
Flag species names in the new dataset that do not occur in the ECD
}
