% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flags.R
\name{flags_for_map}
\alias{flags_for_map}
\title{function for preparing data to be mapped by map_flags}
\usage{
flags_for_map(data, pattern = "^\\\\.", summaryCol = ".summary")
}
\arguments{
\item{data}{new dataset}

\item{pattern}{pattern to index columns that indicate flags}

\item{summaryCol}{column that indicates whether or not an occurence has any flags}
}
\value{
dataframe with new plotNum column used in the curator::map_flags function
}
\description{
function for preparing data to be mapped by map_flags
}
\examples{
flags_for_map(curator::obsData)

}
