% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_types.R
\name{generate_cell_types}
\alias{generate_cell_types}
\title{Get matrices for a specific cell type with specific cell IDs}
\usage{
generate_cell_types(cell_type, sample_list, prefix)
}
\arguments{
\item{cell_type}{A string for which cell type it should be filtering for.}

\item{sample_list}{A list of matrices. Ensure all the matrices have the same number of rows (genes). The genes should be the same for all matrices and in the same order. The names of the columns should be the cell types.}

\item{prefix}{The prefix that your cell IDs should have.}
}
\value{
A sparse filtered matrix.
}
\description{
Get matrices for a specific cell type with specific cell IDs
}
\examples{

set.seed(2024)
sample1 <- matrix(rnorm(20), nrow = 5, ncol = 4,
dimnames = list(paste0("Gene", 1:5),
 c("Ast", "Neu", "Oligo", "Ast")))
sample2 <- matrix(rnorm(20), nrow = 5, ncol = 4,
 dimnames = list(paste0("Gene", 1:5),
  c("Ast", "Ast", "Oligo", "Microglia")))
sample_list <- list(sample1, sample2)

generate_cell_types("Ast", sample_list, "Cohort_name_")
}
