% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_types_w_barcodes.R
\name{generate_cell_types_w_barcodes}
\alias{generate_cell_types_w_barcodes}
\title{Generate cell type specific matrices with barcodes}
\usage{
generate_cell_types_w_barcodes(cell_types, sample_list, cell_IDs)
}
\arguments{
\item{cell_types}{Vector of cell types to process (e.g., c("T_cells", "B_cells")).}

\item{sample_list}{List of matrices where rows are genes and columns are cells. Each matrix should have column names indicating cell types. Genes should be the same (same number and order) for all matrices.}

\item{cell_IDs}{A dataframe that should have two columns: "cell_type" and "cell_ID".}
}
\value{
A list of sparse matrices (one for each cell type in cell_types), with cell IDs as column names and shared row names.
}
\description{
Generate cell type specific matrices with barcodes
}
\examples{

cell_IDs <- data.frame(
cell_type = c("T_cells", "B_cells", "T_cells", "B_cells",
"T_cells", "T_cells", "B_cells", "B_cells"),
cell_ID = c("Cell1", "Cell2", "Cell3", "Cell4",
"Cell5", "Cell6", "Cell7", "Cell8"))

sample1 <- matrix(
c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12),
nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"),
c("Cell1", "Cell2", "Cell3", "Cell4")))

sample2 <- matrix(
c(13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24),
nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"),
c("Cell5", "Cell6", "Cell7", "Cell8")))
sample_list <- list(sample1, sample2)
cell_types <- c("T_cells", "B_cells")
generate_cell_types_w_barcodes(cell_types, sample_list, cell_IDs)
}
