% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_type_file.R
\name{get_cell_type_file}
\alias{get_cell_type_file}
\title{Get cleaned cell type annotations}
\usage{
get_cell_type_file(
  file_path,
  remove_prefix = FALSE,
  remove_suffix = FALSE,
  id_col = NULL,
  type_col = NULL
)
}
\arguments{
\item{file_path}{A string that denotes where the file is.}

\item{remove_prefix}{TRUE or FALSE. Whether to remove the barcodes' prefix.}

\item{remove_suffix}{TRUE or FALSE. Whether to remove the barcodes' suffix.}

\item{id_col}{The name of the barcodes column.}

\item{type_col}{The name of the cell type column.}
}
\value{
A cleaned data table with two columns, cell_id and cell_type.
}
\description{
Get cleaned cell type annotations
}
