% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_continents.R
\name{get_continents}
\alias{get_continents}
\title{Create a continent column for a dataset with a countryCode column}
\usage{
get_continents(data, country = "countryCode", original = "iso3c")
}
\arguments{
\item{data}{dataset with observations}

\item{country}{column name of column with country codes}

\item{original}{type of data in the source column which contains cuntries to identify continent from}
}
\value{
dataset of same type that was passed to it. One column of continents.
}
\description{
Create a continent column for a dataset with a countryCode column
}
\examples{
get_continents(obsData)

}
