% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_continents.R
\name{getContinents}
\alias{getContinents}
\title{Create a continent column for a dataset with a countryCode column}
\usage{
getContinents(data, country = "countryCode")
}
\arguments{
\item{data}{dataset with observations}

\item{country}{column name of countries}
}
\value{
dataset with new continent column
}
\description{
Create a continent column for a dataset with a countryCode column
}
\examples{
getContinents(data)

}
