% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_dge.R
\name{limma_dge}
\alias{limma_dge}
\title{CPMLog normalize and perform DEA}
\usage{
limma_dge(design, PB)
}
\arguments{
\item{design}{To specify the groups of interest. E.g. design <- model.matrix(~ group, data = PB$meta)}

\item{PB}{he output of create_pseudobulk.}
}
\value{
The results of DEA with the limma package.
}
\description{
CPMLog normalize and perform DEA
}
