% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flags.R
\name{map_flags}
\alias{map_flags}
\title{Title}
\usage{
map_flags(
  data,
  longitude = "decimalLongitude",
  latitude = "decimalLatitude",
  plotValues = "plotNum"
)
}
\arguments{
\item{data}{data for plotting}

\item{longitude}{longitude column name}

\item{latitude}{latitude column name}

\item{plotValues}{column that indicates flag state of occurence}
}
\description{
Title
}
\examples{
data = flags_for_map(obsData)
map_flags(data)

}
