% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flags.R
\name{map_obs_and_ECD}
\alias{map_obs_and_ECD}
\title{Function to map ECD observations and new data observations}
\usage{
map_obs_and_ECD(
  data,
  ECD,
  datLongitude = "decimalLongitude",
  datLatitude = "decimalLatitude",
  ECDLongitude = "decimalLongitude",
  ECDLatitude = "decimalLatitude"
)
}
\arguments{
\item{data}{new data for plotting}

\item{ECD}{Expert curated data set for plotting}

\item{datLongitude}{New data column name of column that contains longitude coordinates}

\item{datLatitude}{New data column name of column that contains latitude coordinates}

\item{ECDLongitude}{ECD column name of column that contains longitude coordinates}

\item{ECDLatitude}{ECD column name of column that contains latitude coordinates}
}
\value{
No return value: plots a leaflet map
}
\description{
Function to map ECD observations and new data observations
}
\examples{
map_obs_and_ECD(obsData, ECD)


}
