% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DGE.R
\name{perform_DGE}
\alias{perform_DGE}
\title{TMM Normalize and perform DEA}
\usage{
perform_DGE(PB, design)
}
\arguments{
\item{PB}{The output of create_pseudobulk.}

\item{design}{To specify the groups of interest. E.g. design <- model.matrix(~ group, data = PB$meta)}
}
\value{
The results of DEA with the edgeR package.
}
\description{
TMM Normalize and perform DEA
}
