% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_data.R
\name{rm_flag_data}
\alias{rm_flag_data}
\title{Remove any flagged data from dataset}
\usage{
rm_flag_data(data, sumName = ".summary")
}
\arguments{
\item{data}{name of dataset}

\item{sumName}{name of column that summarizes flags, ie row values in column indicates if data is flagged or not}
}
\value{
dataset of the same type passed to it, but with all flagged data removed
}
\description{
Remove any flagged data from dataset
}
\examples{
rm_flag_data(obsData)

}
