% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCtmat.R
\name{cleanCtmat}
\alias{cleanCtmat}
\title{Filter a matrix based on expression levels}
\usage{
cleanCtmat(ctmat, geneThr = 0.02, sampleThr = 0.05)
}
\arguments{
\item{ctmat}{A numeric matrix where rows represent genes and columns represent samples (e.g., cells).}

\item{geneThr}{A numeric value between 0 and 1 specifying the minimum proportion of cells in which a gene must be expressed to be retained. The default is 0.02.}

\item{sampleThr}{A numeric value between 0 and 1 specifying the minimum rank percentile for the number of genes expressed in a sample to retain that sample. The default is 0.05.}
}
\value{
The filtered matrix.
}
\description{
Filter a matrix based on expression levels
}
\examples{

set.seed(123)
ctmat <- matrix(sample(c(0, 1, 2, 3), 30, replace = TRUE),
nrow = 5, ncol = 6,
dimnames = list(
paste0("Gene", 1:5),
paste0("Cell", 1:6)))
cleanCtmat(ctmat)

}
