% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pseudobulk.R
\name{create_pseudo_bulk}
\alias{create_pseudo_bulk}
\title{Create pseudobulks}
\usage{
create_pseudo_bulk(expr, meta)
}
\arguments{
\item{expr}{Expression matrix where columns are cell IDs and rows are gene names.}

\item{meta}{Dataframe where each row should represent a cell. Cell IDs should be the rownames. It needs to have the columns "patientID", "disorder", "sex", and "age". "disorder" should have the values "yes" or "no".}
}
\value{
A list containing the pseudobulk matrix and its metadata.
}
\description{
Create pseudobulks
}
\examples{
set.seed(123)
expr <- matrix(sample(0:10, 50, replace = TRUE),
nrow = 5, ncol = 10,
dimnames = list(paste0("Gene", 1:5), paste0("Cell", 1:10)))
meta <- data.frame(
patientID = c("P1", "P2", "P3", "P1",
"P1", "P2", "P2", "P3", "P3", "P1"),
disorder = c("yes", "no", "no", "yes",
"yes", "no", "no", "no", "no", "yes"),
sex = c("male", "female", "female", "male",
"male", "female", "female", "female", "female", "male"),
age = c(70, 68, 64, 70, 70, 68, 68, 64, 64, 70),
row.names = paste0("Cell", 1:10)
)

create_pseudo_bulk(expr, meta)

}
