% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flags.R
\name{.map_helper}
\alias{.map_helper}
\title{Helper function to map occurences}
\usage{
.map_helper(data)
}
\arguments{
\item{data}{dataset passed from flags_for_map}
}
\value{
Returns a string indicating whether the observation has no flags, two flags, or which single flag
}
\description{
Helper function to map occurences
}
