% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intoduced_flags.R
\name{flag_introduced}
\alias{flag_introduced}
\title{Title}
\usage{
flag_introduced(data, introCol = "databaseChanges")
}
\arguments{
\item{data}{dataset}

\item{introCol}{name of column that contains introduced species labels}
}
\value{
dataset of the same class passed to it. New column of introduced flags
}
\description{
Title
}
