% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cell_types.R
\name{generate_cell_types}
\alias{generate_cell_types}
\title{Generate cell type specific matrices with unique column names}
\usage{
generate_cell_types(cell_types, sample_list, prefix)
}
\arguments{
\item{cell_types}{Vector of cell types to process (e.g., c("T_cells", "B_cells")).}

\item{sample_list}{List of matrices where rows are genes and columns are cells. Each matrix should have column names indicating cell types. Genes should be the same (same number and order) for all matrices.}

\item{prefix}{to include in the column names of the output.}
}
\value{
A list of sparse matrices (one for each cell type in cell_types), with unique column names and shared row names.
}
\description{
Generate cell type specific matrices with unique column names
}
\examples{

sample1 <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"), c("T_cells", "B_cells")))
sample2 <- matrix(c(7, 8, 9, 10, 11, 12), nrow = 3,
dimnames = list(c("Gene1", "Gene2", "Gene3"), c("T_cells", "B_cells")))
sample_list <- list(sample1, sample2)

cell_types <- c("T_cells", "B_cells")

prefix <- "Sample_"

generate_cell_types(cell_types, sample_list, prefix)
}
