% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flags.R
\name{map_obs_and_ECD}
\alias{map_obs_and_ECD}
\title{Function to map ECD observations and new data observations}
\usage{
map_obs_and_ECD(
  data,
  ECD,
  datLongitude = "decimalLongitude",
  datLatitude = "decimalLatitude",
  ECDLongitude = "decimalLongitude",
  ECDLatitude = "decimalLatitude"
)
}
\arguments{
\item{ECDLatitude}{}
}
\description{
Function to map ECD observations and new data observations
}
