% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obsData}
\alias{obsData}
\title{985 species observations pulled from gbif using get_records_since_date function, then
passed through CoordinateCleaner::clean_coordinates}
\format{
A data frame with 985 rows and 148 variables:
\describe{
\item{scientific name}{scientific name from gbif
...
}
}
\source{
\url{https://www.gbif.org/occurrence/search}
}
\usage{
obsData
}
\description{
A dataset containing biological and location data for 985 gbif obsrvations
}
\keyword{datasets}
